requirements.txt
asgiref==3.8.1
distlib==0.3.9
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
filelock==3.17.0
gunicorn==23.0.0
packaging==24.2
platformdirs==4.3.6
PyJWT==2.9.0
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
tzdata==2025.1
virtualenv==20.29.2
psycopg2-binary
